requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
automat==20.2.0
cffi==1.14.6
channels==3.0.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
django-widget-tweaks==1.4.1
django==3.2.6
djangorestframework==3.6.4
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
olefile==0.44
pillow==8.3.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyopenssl==20.0.1
pytz==2017.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
turbo-django==0.2.0
twisted[tls]==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
zope.interface==5.4.0
